+++ Solar-System-2.0/Engine/GUI/Menu.cpp
@@ -63,7 +63,7 @@ void Menu::render(int width, int height)
     ImVec2 frame_padding_save = style.FramePadding;
 
     window_flags |= ImGuiWindowFlags_NoResize;
-    ImGui::SetNextWindowPos(ImVec2(width / 2 - 200, height / 2 - 200));
+    ImGui::SetNextWindowPos(ImVec2(((float)width / 2.f) - 200.f, ((float)height / 2.f) - 200.f));
     ImGui::SetNextWindowSize(ImVec2(400, 400));
